kafka cleanup and include of mllib
